Revert "extend the base service class and move all authorize logic into the service layers"
This reverts commit 9895381

+++ src/Http/Controllers/Api/DiscussionController.php
@@ -37,6 +37,8 @@ class DiscussionController extends Controller
      */
     public function store(StoreDiscussionRequest $request)
     {
+        $this->authorize('create', Discussion::class);
+
         $group = Group::findOrFail($request->group_id);
 
         $discussion = $this->discussionService->make($request, $group);
@@ -52,6 +54,8 @@ class DiscussionController extends Controller
      */
     public function update(UpdateDiscussionRequest $request, Discussion $discussion)
     {
+        $this->authorize('update', $discussion);
+
         $discussion = $this->discussionService->update($request, $discussion);
 
         return response()->json($discussion->fresh());
@@ -65,6 +69,8 @@ class DiscussionController extends Controller
      */
     public function delete(Request $request, Discussion $discussion)
     {
+        $this->authorize('delete', $discussion);
+
         $this->discussionService->delete($discussion);
 
         return response()->json('OK');
@@ -79,6 +85,8 @@ class DiscussionController extends Controller
      */
     public function lock(Request $request, Discussion $discussion)
     {
+        $this->authorize('lock', $discussion);
+
         $this->discussionService->updateState($discussion, 'lock');
 
         return response()->json('OK');
@@ -93,6 +101,8 @@ class DiscussionController extends Controller
      */
     public function unlock(Request $request, Discussion $discussion)
     {
+        $this->authorize('unlock', $discussion);
+
         $this->discussionService->updateState($discussion, 'unlock');
 
         return response()->json('OK');
@@ -107,6 +117,8 @@ class DiscussionController extends Controller
      */
     public function pin(Request $request, Discussion $discussion)
     {
+        $this->authorize('pin', $discussion);
+
         $this->discussionService->updateState($discussion, 'pin');
 
         return response()->json('OK');
@@ -121,6 +133,8 @@ class DiscussionController extends Controller
      */
     public function unpin(Request $request, Discussion $discussion)
     {
+        $this->authorize('unpin', $discussion);
+
         $this->discussionService->updateState($discussion, 'unpin');
 
         return response()->json('OK');
